태스크 - 사용자 매핑

+++ src/main/java/io/poten13/deepfocus/domain/task/controller/UpdateTaskRequest.java
@@ -17,9 +17,9 @@ public class UpdateTaskRequest {
     @Schema(description = "태스크 제목", example = "책 읽기")
     private String title;
     @Schema(description = "태스크 시작 시간 (unixTimestamp)", example = "1699224390")
-    private int startTime;
+    private long startTime;
     @Schema(description = "태스크 수행 시간 (단위: 분)", example = "30")
-    private int spanMinute;
+    private long spanMinute;
     @Schema(description = "태스크 의 하위 태스크 목록")
     List<UpdateSubTaskRequest> subTasks;
 }
